+++ C-Assignments/Assignment -1/Product.c
@@ -1,4 +1,4 @@
-long int product(int x, int y) {
+long int product(long int x, long int y) {
 	extern long int mul;
 	mul = x * y;
 	return mul;
